+++ src/bin/vip-dev-env-import-sql.js
@@ -9,17 +9,19 @@
  * External dependencies
  */
 import fs from 'fs';
+import chalk from 'chalk';
 
 /**
  * Internal dependencies
  */
-import { trackEvent } from 'lib/tracker';
+import { trackEvent } from '../lib/tracker';
 import command from '../lib/cli/command';
 import { getEnvironmentName, getEnvTrackingInfo, handleCLIException, promptForBoolean, validateDependencies } from '../lib/dev-environment/dev-environment-cli';
-import { exec, resolveImportPath } from '../lib/dev-environment/dev-environment-core';
+import { exec, getEnvironmentPath, resolveImportPath } from '../lib/dev-environment/dev-environment-core';
 import { DEV_ENVIRONMENT_FULL_COMMAND } from '../lib/constants/dev-environment';
 import { validate } from '../lib/validations/sql';
-import { bootstrapLando } from '../lib/dev-environment/dev-environment-lando';
+import { bootstrapLando, isEnvUp } from '../lib/dev-environment/dev-environment-lando';
+import UserError from '../lib/user-error';
 
 const examples = [
 	{
@@ -60,9 +62,13 @@ command( {
 		await trackEvent( 'dev_env_import_sql_command_execute', trackingInfo );
 
 		try {
-			const { resolvedPath, inContainerPath } = await resolveImportPath( slug, fileName, searchReplace, inPlace );
+			const resolvedPath = await resolveImportPath( slug, fileName, searchReplace, inPlace );
 
 			if ( ! opt.skipValidate ) {
+				if ( ! isEnvUp( lando, getEnvironmentPath( slug ) ) ) {
+					throw new UserError( 'Environment needs to be started first' );
+				}
+
 				const expectedDomain = `${ slug }.vipdev.lndo.site`;
 				await validate( resolvedPath, {
 					isImport: false,
@@ -71,10 +77,28 @@ command( {
 				} );
 			}
 
-			const importArg = [ 'wp', 'db', 'import', inContainerPath ];
-			await exec( lando, slug, importArg );
+			const fd = await fs.promises.open( resolvedPath, 'r' );
+			const importArg = [ 'db', '--disable-auto-rehash' ];
+			const origIsTTY = process.stdin.isTTY;
 
-			fs.unlinkSync( resolvedPath );
+			try {
+				/**
+				 * When stdin is a TTY, Lando passes the `--tty` flag to Docker.
+				 * This breaks our code when we pass the stream as stdin to Docker.
+				 * exec() then fails with "the input device is not a TTY".
+				 *
+				 * Therefore, for the things to work, we have to pretend that stdin is not a TTY :-)
+				 */
+				process.stdin.isTTY = false;
+				await exec( lando, slug, importArg, { stdio: [ fd, 'pipe', 'pipe' ] } );
+				console.log( `${ chalk.green.bold( 'Success:' ) } Database imported.` );
+			} finally {
+				process.stdin.isTTY = origIsTTY;
+			}
+
+			if ( searchReplace && searchReplace.length && ! inPlace ) {
+				fs.unlinkSync( resolvedPath );
+			}
 
 			const cacheArg = [ 'wp', 'cache', 'flush' ];
 			await exec( lando, slug, cacheArg );
@@ -93,6 +117,7 @@ command( {
 			await exec( lando, slug, addUserArg );
 			await trackEvent( 'dev_env_import_sql_command_success', trackingInfo );
 		} catch ( error ) {
-			handleCLIException( error, 'dev_env_import_sql_command_error', trackingInfo );
+			await handleCLIException( error, 'dev_env_import_sql_command_error', trackingInfo );
+			process.exitCode = 1;
 		}
 	} );