+++ src/bin/vip-dev-env-info.js
@@ -61,6 +61,7 @@ command()
 			}
 			await trackEvent( 'dev_env_info_command_success', trackingInfo );
 		} catch ( error ) {
-			handleCLIException( error, 'dev_env_info_command_error', trackingInfo );
+			await handleCLIException( error, 'dev_env_info_command_error', trackingInfo );
+			process.exitCode = 1;
 		}
 	} );