+++ src/bin/vip-dev-env-list.js
@@ -31,6 +31,7 @@ command()
 	.examples( examples )
 	.argv( process.argv, async () => {
 		const lando = await bootstrapLando();
+		lando.events.constructor.prototype.setMaxListeners( 1024 );
 		await validateDependencies( lando, '' );
 
 		const trackingInfo = { all: true };
@@ -40,6 +41,7 @@ command()
 			await printAllEnvironmentsInfo( lando, {} );
 			await trackEvent( 'dev_env_list_command_success', trackingInfo );
 		} catch ( error ) {
-			handleCLIException( error, 'dev_env_list_command_error', trackingInfo );
+			await handleCLIException( error, 'dev_env_list_command_error', trackingInfo );
+			process.exitCode = 1;
 		}
 	} );