+++ src/lib/dev-environment/dev-environment-core.js
@@ -58,11 +58,6 @@ type StartEnvironmentOptions = {
 	skipWpVersionsCheck: boolean
 };
 
-type SQLImportPaths = {
-	resolvedPath: string,
-	inContainerPath: string
-}
-
 type WordPressTag = {
 	ref: string;
 	tag: string;
@@ -173,6 +168,11 @@ function preProcessInstanceData( instanceData: InstanceData ): InstanceData {
 		newInstanceData.xdebugConfig = '';
 	}
 
+	// Mailhog migration
+	if ( ! newInstanceData.mailhog ) {
+		newInstanceData.mailhog = false;
+	}
+
 	return newInstanceData;
 }
 
@@ -182,8 +182,7 @@ export async function destroyEnvironment( lando: Lando, slug: string, removeFile
 
 	debug( 'Instance path for', slug, 'is:', instancePath );
 
-	const environmentExists = fs.existsSync( instancePath );
-
+	const environmentExists = await doesEnvironmentExist( instancePath );
 	if ( ! environmentExists ) {
 		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
 	}
@@ -213,8 +212,16 @@ export async function printAllEnvironmentsInfo( lando: Lando, options: PrintOpti
 
 	console.log( 'Found ' + chalk.bold( allEnvNames.length ) + ' environments' + ( allEnvNames.length ? ':' : '.' ) );
 	for ( const envName of allEnvNames ) {
-		console.log( '\n' );
-		await printEnvironmentInfo( lando, envName, options );
+		try {
+			console.log( '\n' );
+			await printEnvironmentInfo( lando, envName, options );
+		} catch ( error ) {
+			if ( error instanceof UserError ) {
+				console.warn( '\nWARNING: "%s" is not a valid environment\n', envName );
+			} else {
+				throw error;
+			}
+		}
 	}
 }
 
@@ -232,10 +239,9 @@ export async function printEnvironmentInfo( lando: Lando, slug: string, options:
 
 	debug( 'Instance path for', slug, 'is:', instancePath );
 
-	const environmentExists = fs.existsSync( instancePath );
-
+	const environmentExists = await doesEnvironmentExist( instancePath );
 	if ( ! environmentExists ) {
-		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
+		throw new UserError( DEV_ENVIRONMENT_NOT_FOUND );
 	}
 
 	const appInfo = await landoInfo( lando, instancePath );
@@ -255,34 +261,26 @@ export async function printEnvironmentInfo( lando: Lando, slug: string, options:
 	printTable( appInfo );
 }
 
-export async function exec( lando: Lando, slug: string, args: Array<string>, options: any = {} ) {
+export function exec( lando: Lando, slug: string, args: Array<string>, options: any = {} ): Promise<*> {
 	debug( 'Will run a wp command on env', slug, 'with args', args, ' and options', options );
 
 	const instancePath = getEnvironmentPath( slug );
 
 	debug( 'Instance path for', slug, 'is:', instancePath );
 
-	const environmentExists = fs.existsSync( instancePath );
-
-	if ( ! environmentExists ) {
-		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
-	}
-
-	const command = args.shift();
-
-	const commandArgs = [ ...args ];
-
-	await landoExec( lando, instancePath, command, commandArgs, options );
+	const [ command, ...commandArgs ] = args;
+	return landoExec( lando, instancePath, command, commandArgs, options );
 }
 
-export function doesEnvironmentExist( slug: string ): boolean {
-	debug( 'Will check for environment', slug );
-
-	const instancePath = getEnvironmentPath( slug );
-
-	debug( 'Instance path for', slug, 'is:', instancePath );
-
-	return fs.existsSync( instancePath );
+export async function doesEnvironmentExist( instancePath: string ): Promise<boolean> {
+	debug( 'Will check for environment at', instancePath );
+	const file = path.join( instancePath, instanceDataFileName );
+	try {
+		const stats = await fs.promises.stat( file );
+		return stats.isFile();
+	} catch ( err ) {
+		return false;
+	}
 }
 
 export function readEnvironmentData( slug: string ): InstanceData {
@@ -456,30 +454,19 @@ export async function getApplicationInformation( appId: number, envType: string 
 	return appData;
 }
 
-export async function resolveImportPath( slug: string, fileName: string, searchReplace: string | string[], inPlace: boolean ): Promise<SQLImportPaths> {
+export async function resolveImportPath( slug: string, fileName: string, searchReplace: string | string[], inPlace: boolean ): Promise<string> {
 	debug( `Will try to resolve path - ${ fileName }` );
 	let resolvedPath = resolvePath( fileName );
-
-	const instancePath = getEnvironmentPath( slug );
-
-	debug( `Instance path for ${ slug } is ${ instancePath }` );
-
-	const environmentExists = fs.existsSync( instancePath );
-
-	if ( ! environmentExists ) {
-		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
-	}
 
 	debug( `Filename ${ fileName } resolved to ${ resolvedPath }` );
 
 	if ( ! fs.existsSync( resolvedPath ) ) {
 		throw new UserError( `The provided file ${ resolvedPath } does not exist or it is not valid (see "--help" for examples)` );
 	}
+
 	if ( fs.lstatSync( resolvedPath ).isDirectory() ) {
 		throw new UserError( `The provided file ${ resolvedPath } is a directory. Please point to a sql file.` );
 	}
-
-	let baseName: string;
 
 	// Run Search and Replace if the --search-replace flag was provided
 	if ( searchReplace && searchReplace.length ) {
@@ -494,22 +481,9 @@ export async function resolveImportPath( slug: string, fileName: string, searchR
 		}
 
 		resolvedPath = outputFileName;
-		baseName = path.basename( outputFileName );
-	} else {
-		baseName = path.basename( resolvedPath );
 	}
 
-	const targetPath = path.join( instancePath, baseName );
-	const inContainerPath = `/app/${ baseName }`;
-	debug( `Copying ${ resolvedPath } to ${ targetPath }` );
-	fs.copyFileSync( resolvedPath, targetPath, fs.constants.COPYFILE_FICLONE );
-	debug( `Copied ${ resolvedPath } to ${ targetPath }` );
-
-	debug( `Import file path ${ resolvedPath } will be mapped to ${ inContainerPath }` );
-	return {
-		resolvedPath: targetPath,
-		inContainerPath,
-	};
+	return resolvedPath;
 }
 
 export async function importMediaPath( slug: string, filePath: string ) {
@@ -517,6 +491,11 @@ export async function importMediaPath( slug: string, filePath: string ) {
 
 	if ( ! fs.existsSync( resolvedPath ) || ! fs.lstatSync( resolvedPath ).isDirectory() ) {
 		throw new Error( 'The provided path does not exist or it is not valid (see "--help" for examples)' );
+	}
+
+	const environmentPath = getEnvironmentPath( slug );
+	if ( ! await doesEnvironmentExist( environmentPath ) ) {
+		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
 	}
 
 	const files = fs.readdirSync( resolvedPath );
@@ -532,7 +511,6 @@ export async function importMediaPath( slug: string, filePath: string ) {
 		}
 	}
 
-	const environmentPath = getEnvironmentPath( slug );
 	const uploadsPath = path.join( environmentPath, uploadPathString );
 
 	console.log( `${ chalk.yellow( '-' ) } Started copying files` );
@@ -566,7 +544,8 @@ async function updateWordPressImage( slug: string ): Promise<boolean> {
 		} else {
 			message = `An error prevented reading the configuration of: ${ slug }\n\n ${ error }`;
 		}
-		handleCLIException( new Error( message ) );
+
+		await handleCLIException( new Error( message ) );
 		return false;
 	}