+++ src/lib/dev-environment/dev-environment-lando.js
@@ -14,13 +14,15 @@ import landoUtils from 'lando/plugins/lando-core/lib/utils';
 import landoBuildTask from 'lando/plugins/lando-tooling/lib/build';
 import chalk from 'chalk';
 import App from 'lando/lib/app';
-import UserError from '../user-error';
 import dns from 'dns';
 
 /**
  * Internal dependencies
  */
-import { readEnvironmentData, writeEnvironmentData } from './dev-environment-core';
+import { doesEnvironmentExist, readEnvironmentData, writeEnvironmentData } from './dev-environment-core';
+import { DEV_ENVIRONMENT_NOT_FOUND } from '../constants/dev-environment';
+import UserError from '../user-error';
+
 /**
  * This file will hold all the interactions with lando library
  */
@@ -44,9 +46,11 @@ function getLandoUserConfigurationRoot() {
  * @returns {object} Lando configuration
  */
 function getLandoConfig() {
-	const landoPath = path.join( __dirname, '..', '..', '..', 'node_modules', 'lando' );
+	const nodeModulesPath = path.join( __dirname, '..', '..', '..', 'node_modules' );
+	const landoPath = path.join( nodeModulesPath, 'lando' );
+	const atLandoPath = path.join( nodeModulesPath, '@lando' );
 
-	debug( `Getting lando config, using path '${ landoPath }' for plugins` );
+	debug( `Getting lando config, using paths '${ landoPath }' and '${ atLandoPath }' for plugins` );
 
 	const isLandoDebugSelected = ( process.env.DEBUG || '' ).includes( DEBUG_KEY );
 	const isAllDebugSelected = process.env.DEBUG === '*';
@@ -60,14 +64,77 @@ function getLandoConfig() {
 		pluginDirs: [
 			landoPath,
 			{
-				path: path.join( landoPath, 'integrations' ),
+				path: atLandoPath,
 				subdir: '.',
+				namespace: '@lando',
 			},
+		],
+		disablePlugins: [
+			// Plugins we need:
+			// '@lando/compose',
+			// '@lando/mailhog',
+			// '@lando/memcached',
+			// '@lando/phpmyadmin',
+			// The rest we don't need
+			'@lando/acquia',
+			'@lando/apache',
+			'@lando/argv',
+			'@lando/backdrop',
+			'@lando/dotnet',
+			'@lando/drupal',
+			'@lando/elasticsearch',
+			'@lando/go',
+			'@lando/joomla',
+			'@lando/lagoon',
+			'@lando/lamp',
+			'@lando/laravel',
+			'@lando/lemp',
+			'@lando/mariadb',
+			'@lando/mean',
+			'@lando/mongo',
+			'@lando/mssql',
+			'@lando/mysql',
+			'@lando/nginx',
+			'@lando/node',
+			'@lando/pantheon',
+			'@lando/php',
+			'@lando/platformsh',
+			'@lando/postgres',
+			'@lando/python',
+			'@lando/redis',
+			'@lando/ruby',
+			'@lando/solr',
+			'@lando/symfony',
+			'@lando/tomcat',
+			'@lando/varnish',
+			'@lando/wordpress',
 		],
 		proxyName: 'vip-dev-env-proxy',
 		userConfRoot: getLandoUserConfigurationRoot(),
 		home: '',
 	};
+}
+
+const appMap: Map<string, App> = new Map();
+
+async function getLandoApplication( lando: Lando, instancePath: string ): Promise<App> {
+	if ( appMap.has( instancePath ) ) {
+		return Promise.resolve( appMap.get( instancePath ) );
+	}
+
+	if ( ! await doesEnvironmentExist( instancePath ) ) {
+		throw new Error( DEV_ENVIRONMENT_NOT_FOUND );
+	}
+
+	const app = lando.getApp( instancePath );
+	addHooks( app, lando );
+	appMap.set( instancePath, app );
+
+	if ( ! app.initialized ) {
+		await app.init();
+	}
+
+	return app;
 }
 
 export async function bootstrapLando(): Promise<Lando> {
@@ -79,31 +146,22 @@ export async function bootstrapLando(): Promise<Lando> {
 export async function landoStart( lando: Lando, instancePath: string ) {
 	debug( 'Will start lando app on path:', instancePath );
 
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	addHooks( app, lando );
-
+	const app = await getLandoApplication( lando, instancePath );
 	await app.start();
 }
 
 export async function landoRebuild( lando: Lando, instancePath: string ) {
 	debug( 'Will rebuild lando app on path:', instancePath );
 
-	const app = lando.getApp( instancePath );
-	await app.init();
-
+	const app = await getLandoApplication( lando, instancePath );
 	await ensureNoOrphantProxyContainer( lando );
-
-	addHooks( app, lando );
-
 	await app.rebuild();
 }
 
 function addHooks( app: App, lando: Lando ) {
 	app.events.on( 'post-start', 1, () => healthcheckHook( app, lando ) );
 
-	lando.events.on( 'pre-engine-build', 5, async data => {
+	lando.events.once( 'pre-engine-build', async data => {
 		const instanceData = readEnvironmentData( app._name );
 
 		let registryResolvable = false;
@@ -130,21 +188,25 @@ function addHooks( app: App, lando: Lando ) {
 
 const healthChecks = {
 	database: 'mysql -uroot --silent --execute "SHOW DATABASES;"',
-	'vip-search': "curl -s --noproxy '*' -XGET localhost:9200",
+	elasticsearch: "curl -s --noproxy '*' -XGET localhost:9200",
 	php: '[[ -f /wp/wp-includes/pomo/mo.php ]]',
 };
 
 async function healthcheckHook( app: App, lando: Lando ) {
+	const now = new Date();
 	try {
 		await lando.Promise.retry( async () => {
 			const list = await lando.engine.list( { project: app.project } );
 
 			const notHealthyContainers = [];
+			const checkPromises = [];
+			const containerOrder = [];
 			for ( const container of list ) {
 				if ( healthChecks[ container.service ] ) {
-					try {
-						debug( `Testing ${ container.service }: ${ healthChecks[ container.service ] }` );
-						await app.engine.run( {
+					debug( `Testing ${ container.service }: ${ healthChecks[ container.service ] }` );
+					containerOrder.push( container );
+					checkPromises.push(
+						app.engine.run( {
 							id: container.id,
 							cmd: healthChecks[ container.service ],
 							compose: app.compose,
@@ -155,56 +217,55 @@ async function healthcheckHook( app: App, lando: Lando ) {
 								cstdio: 'pipe',
 								services: [ container.service ],
 							},
-						} );
-					} catch ( exception ) {
-						debug( `${ container.service } Health check failed` );
-						notHealthyContainers.push( container );
-					}
+						} )
+					);
 				}
 			}
 
-			if ( notHealthyContainers.length ) {
-				for ( const container of notHealthyContainers ) {
-					console.log( `Waiting for service ${ container.service } ...` );
+			const results = await Promise.allSettled( checkPromises );
+			results.forEach( ( result, index ) => {
+				if ( result.status === 'rejected' ) {
+					debug( `${ containerOrder[ index ].service } Health check failed` );
+					notHealthyContainers.push( containerOrder[ index ] );
 				}
+			} );
+
+			if ( notHealthyContainers.length ) {
+				notHealthyContainers.forEach( container => console.log( `Waiting for service ${ container.service } ...` ) );
 				return Promise.reject( notHealthyContainers );
 			}
 		}, { max: 20, backoff: 1000 } );
 	} catch ( containersWithFailingHealthCheck ) {
-		for ( const container of containersWithFailingHealthCheck ) {
-			console.log( chalk.yellow( 'WARNING:' ) + ` Service ${ container.service } failed healthcheck` );
-		}
+		containersWithFailingHealthCheck.forEach( container => console.log( chalk.yellow( 'WARNING:' ) + ` Service ${ container.service } failed healthcheck` ) );
 	}
+
+	const duration = new Date().getTime() - now.getTime();
+	debug( `Healthcheck completed in ${ duration }ms` );
 }
 
 export async function landoStop( lando: Lando, instancePath: string ) {
 	debug( 'Will stop lando app on path:', instancePath );
 
-	const app = lando.getApp( instancePath );
-	await app.init();
-
+	const app = await getLandoApplication( lando, instancePath );
 	await app.stop();
 }
 
 export async function landoDestroy( lando: Lando, instancePath: string ) {
 	debug( 'Will destroy lando app on path:', instancePath );
 
-	const app = lando.getApp( instancePath );
-	await app.init();
-
+	const app = await getLandoApplication( lando, instancePath );
 	await app.destroy();
 }
 
 export async function landoInfo( lando: Lando, instancePath: string ) {
-	const app = lando.getApp( instancePath );
-	await app.init();
+	const app = await getLandoApplication( lando, instancePath );
 
 	let appInfo = landoUtils.startTable( app );
 
 	const reachableServices = app.info.filter( service => service.urls.length );
 	reachableServices.forEach( service => appInfo[ `${ service.service } urls` ] = service.urls );
 
-	const isUp = await isEnvUp( app );
+	const isUp = await isEnvUp( lando, instancePath );
 	const frontEndUrl = app.info
 		.find( service => 'nginx' === service.service )
 		?.urls[ 0 ];
@@ -237,7 +298,7 @@ export async function landoInfo( lando: Lando, instancePath: string ) {
 
 const extraServiceDisplayConfiguration = [
 	{
-		name: 'vip-search',
+		name: 'elasticsearch',
 		label: 'enterprise search',
 		protocol: 'http',
 	},
@@ -279,50 +340,57 @@ async function getExtraServicesConnections( lando, app ) {
 	return extraServices;
 }
 
-async function isEnvUp( app ) {
+export async function isEnvUp( lando: Lando, instancePath: string ): Promise<boolean> {
+	const now = new Date();
+	const app = await getLandoApplication( lando, instancePath );
+
 	const reachableServices = app.info.filter( service => service.urls.length );
 	const urls = reachableServices.map( service => service.urls ).flat();
 
 	const scanResult = await app.scanUrls( urls, { max: 1 } );
+	const duration = new Date().getTime() - now.getTime();
+	debug( 'isEnvUp took %d ms', duration );
+
 	// If all the URLs are reachable then the app is considered 'up'
 	return scanResult?.length && scanResult.filter( result => result.status ).length === scanResult.length;
 }
 
 export async function landoExec( lando: Lando, instancePath: string, toolName: string, args: Array<string>, options: any ) {
-	const app = lando.getApp( instancePath );
-	await app.init();
-
-	if ( ! options.force ) {
-		const isUp = await isEnvUp( app );
-
-		if ( ! isUp ) {
-			throw new UserError( 'Environment needs to be started before running wp command' );
-		}
-	}
+	const app = await getLandoApplication( lando, instancePath );
 
 	const tool = app.config.tooling[ toolName ];
 	if ( ! tool ) {
-		throw new Error( `${ toolName } is not a known lando task` );
+		throw new UserError( `${ toolName } is not a known lando task` );
 	}
 
-	/*
-	 lando is looking in both passed args and process.argv so we need to do a bit of hack to fake process.argv
-	 so that lando doesn't try to interpret args not meant for wp.
+	const savedArgv = process.argv;
+	try {
+		/*
+			lando is looking in both passed args and process.argv so we need to do a bit of hack to fake process.argv
+			so that lando doesn't try to interpret args not meant for wp.
 
-	 Lando drops first 3 args (<node> <lando> <command>) from process.argv and process rest, so we will fake 3 args + the real args
-	*/
-	process.argv = [ '0', '1', '3' ].concat( args );
+			Lando drops first 3 args (<node> <lando> <command>) from process.argv and process rest, so we will fake 3 args + the real args
+		*/
+		process.argv = [ '0', '1', '3' ].concat( args );
 
-	tool.app = app;
-	tool.name = toolName;
+		tool.app = app;
+		tool.name = toolName;
 
-	const task = landoBuildTask( tool, lando );
+		if ( options.stdio ) {
+			tool.stdio = options.stdio;
+		}
 
-	const argv = {
-		_: args, // eslint-disable-line
-	};
+		const task = landoBuildTask( tool, lando );
 
-	await task.run( argv );
+		const argv = {
+			// eslint-disable-next-line id-length
+			_: args,
+		};
+
+		await task.run( argv );
+	} finally {
+		process.argv = savedArgv;
+	}
 }
 
 /**
@@ -335,7 +403,7 @@ export async function landoExec( lando: Lando, instancePath: string, toolName: s
  *
  * @param {object} lando Bootstrapped Lando object
  */
-async function ensureNoOrphantProxyContainer( lando ) {
+async function ensureNoOrphantProxyContainer( lando: Lando ) {
 	const proxyContainerName = lando.config.proxyContainer;
 
 	const docker = lando.engine.docker;